Refactor graph component
