+++ air-quality-app/client/src/App.tsx
@@ -1,36 +1,9 @@
 import React, {useEffect, useState} from 'react';
 import './App.css';
-
-import { GraphsContainer } from "./components/GraphsContainer";
-import { PeriodPicker } from "./components/PeriodPicker";
-import { GraphResponseData, TimePeriod } from "./types";
-import { getData} from "./api/api";
+import { GraphComponent } from "./components/graphs/GraphComponent";
 import { MapSelector } from './components/MapSelector';
 
 function App() {
-  let [graphData, setGraphData] = useState<GraphResponseData>();
-  let [timePeriod, setTimePeriod] = useState<TimePeriod>(["2021-02-02 11:00:00", "2021-02-02 12:00:00"]);
-  let [isLoading, setIsLoading] = useState(false);
-  let [urlSuffix, setUrlSuffix] = useState<string>("Average");
-
-  useEffect(() => {
-    setIsLoading(true);
-    getData(timePeriod, urlSuffix).then((response) => {
-      let responseData = JSON.parse(response)
-      setGraphData(responseData)
-      setIsLoading(false);
-    })
-    
-  }, [timePeriod])
-
-  const handlePeriodChange = (period: TimePeriod = ["2021-02-02 11:00:00", "2021-02-02 12:00:00"]) => {
-    setTimePeriod(period)
-  }
-
-  const handleUrlChange = (url: string) => {
-    setUrlSuffix(url)
-  }
-
   //rendering logic
   return (
     <div className="App">
@@ -42,15 +15,7 @@ function App() {
           <h3> Tytul tego wszystkiego</h3>
           <p>Lorem ipsum i takie tam</p>
         </div>
-        <PeriodPicker onChange={(period) => {handlePeriodChange(period)}} 
-                      isLoading={isLoading} 
-                      onStationChange={(stationName) => handleUrlChange(stationName)} />
-        <div className='Graphs'>
-          <GraphsContainer text="CO" data={graphData?.co}/>
-          <GraphsContainer text='NO2' data={graphData?.no}/>
-          <GraphsContainer text='PM10' data={graphData?.pm}/>
-        </div>
-		
+        <GraphComponent/>
         <MapSelector />
         <div className='Conclusions'>
           Ziemia i środowisko zmagają się z poważnym problemem zwiększającego się zanieczyszczenia powietrza, wody oraz gleby.  Zanieczyszczenia te są spowodowane w większości działalnością człowieka i mogą mieć fatalne skutki dla flory, fauny i ludzi. Jednymi z najgroźniejszych cząsteczek są tlenki węgla, dwutlenki azotu oraz różne inne cząsteczki zawieszone w powietrzu o średnicy nie większej niż 10 mikrometrów (określane nazwą PM10), pochodzące najczęściej z spalania paliw stałych i ciekłych. <br></br><br></br>
