+++ air-quality-app/client/src/components/graphs/GraphsContainer.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { GraphRenderer } from "./GraphRenderer";
-import { GraphData } from "../types";
+import { GraphData } from "../../types";
 import '../styles/GraphContainerStyling.css'
 
 type Props = {text: string, data?: GraphData}
